requirements.txt
# AlignTune - Complete Dependencies
torch>=2.7.1
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0

# Core Python dependencies
numpy>=1.21.0
pyyaml>=6.0
huggingface-hub>=0.17.0
tqdm>=4.64.0
psutil>=5.9.0

# Data processing
pandas>=1.3.0
scikit-learn>=1.0.0

# Utilities
click>=8.0.0
rich>=12.0.0
typer>=0.7.0

# Reinforcement Learning
trl==0.23.0
vllm

# Unsloth acceleration
unsloth==2025.11.1

# Logging and monitoring
wandb>=0.15.0
tensorboard>=2.12.0
matplotlib>=3.5.0
seaborn>=0.11.0
xformers

# Evaluation metrics
evaluate>=0.4.0
rouge-score>=0.1.2
sacrebleu>=2.3.0
lm-eval>=0.4.0
nltk>=3.8.0
# Optional evaluation metrics (for enhanced SFT evaluation - install separately if needed)
bert-score>=0.3.13  # For BERTScore metrics
sentence-transformers>=2.2.0  # For semantic similarity metrics
codebleu>=0.1.0  # For CodeBLEU metrics (code generation tasks)

# Device management (Linux/Windows only)
nvitop>=1.1.0; platform_system!='Darwin'
nvidia-ml-py>=11.495.46; platform_system!='Darwin'
gpustat>=1.1.1

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
pytest-mock>=3.10.0
pytest-timeout>=2.1.0
hypothesis>=6.70.0

# Code quality tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.2.0
myst-parser>=0.18.0
sphinx-autodoc-typehints>=1.19.0